+++ src/pages/home/index.js
@@ -27,7 +27,7 @@ import {
   Voiceflow,
   Wolfram
 } from 'assets/sponsors';
-import styles from './index.module.css';
+import styles from './home.module.css';
 import { Link } from 'react-router-dom'; 
 
 function Landing() {
